refactor: split tui_context.rs into tui_context/ module
CQ-4: The 1367-line TUI god object is split into a module directory
with clear cohesion boundaries:

- tui_context/mod.rs (559 lines) — struct, constructor, lifecycle,
  command dispatch
- tui_context/events.rs (408 lines) — keyboard, mouse, paste,
  clipboard, history input handling
- tui_context/menus.rs (419 lines) — dropdown pickers, menu nav,
  wizard state machines

Cross-file methods promoted to pub(crate). Regression test updated
to point at mod.rs. Resolved the TODO at line 43 about splitting.

No logic changes — pure structural refactor.
